[YS-513] refactor: EmailBadge 컴포넌트 분리
