+++ src/app/join/mobile/Participant/ContactEmailStep/ContactEmailStep.tsx
@@ -1,22 +1,15 @@
-import Image from 'next/image';
 import { FormProvider, useFormContext } from 'react-hook-form';
 
 import ServiceAgreeBottomSheet from '../../components/ServiceAgreeBottomSheet/ServiceAgreeBottomSheet';
 import TitleSection from '../../components/TitleSection/TitleSection';
-import { email, emailWrapper, mainContentLayout } from '../../page.css';
+import { mainContentLayout } from '../../page.css';
 import NextButton from './NextButton/NextButton';
 
-import Google from '@/assets/images/google.svg';
-import Naver from '@/assets/images/naver.svg';
 import ContactEmailInput from '@/components/ContactEmailInput/ContactEmailInput';
+import EmailBadge from '@/components/EmailBadge/EmailBadge';
 import useOverlay from '@/hooks/useOverlay';
 import { ParticipantJoinSchemaType } from '@/schema/join/ParticipantJoinSchema';
 import { LoginProvider } from '@/types/user';
-
-const logoMap = {
-  NAVER: Naver,
-  GOOGLE: Google,
-};
 
 interface ContactEmailStepProps {
   onNext: () => void;
@@ -50,12 +43,7 @@ const ContactEmailStep = ({ onNext, provider, oauthEmail }: ContactEmailStepProp
       <TitleSection
         title="연락 받을 이메일을 입력해 주세요"
         description="로그인 아이디와 달라도 괜찮아요"
-        emailBadge={
-          <div className={emailWrapper}>
-            <Image src={logoMap[provider]} alt="로고" width={24} height={24} />
-            <span className={email}>{oauthEmail}</span>
-          </div>
-        }
+        emailBadge={<EmailBadge provider={provider} oauthEmail={oauthEmail} />}
       />
       <ContactEmailInput<ParticipantJoinSchemaType>
         contactEmailField="contactEmail"
